Inhibit block enhancements.

+++ workspace/Block/src/com/ils/block/Compare.java
@@ -73,7 +73,7 @@ public class Compare extends AbstractProcessBlock implements ProcessBlock {
 	private void initialize() {		
 		setName("Compare");
 		// Define the time for "coalescing" inputs ~ msec
-		BlockProperty oprop = new BlockProperty(BlockConstants.BLOCK_PROPERTY_OFFSET,new Double(offset),PropertyType.INTEGER,true);
+		BlockProperty oprop = new BlockProperty(BlockConstants.BLOCK_PROPERTY_OFFSET,new Double(offset),PropertyType.DOUBLE,true);
 		setProperty(BlockConstants.BLOCK_PROPERTY_OFFSET, oprop);
 		BlockProperty synch = new BlockProperty(BlockConstants.BLOCK_PROPERTY_SYNC_INTERVAL,new Double(synchInterval),PropertyType.TIME,true);
 		setProperty(BlockConstants.BLOCK_PROPERTY_SYNC_INTERVAL, synch);
